Conversation Context & Provider
Created conversations context and provider to give real-time updates without reloading UI
